Refactor cloud base height: move to cloud.ts and use 124.7 constant
